pyproject.toml
[tool.poetry]
name = "vin-decoder-proj"
version = "0.1.0"
description = ""
authors = ["Ivan <ashikhin1998@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
postgres = "^4.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
vininfo = "^1.7.0"
drf-spectacular = "^0.26.5"
django-sslserver = "^0.22"
pytest = "^7.4.3"
pytest-django = "^4.7.0"
django-cors-headers = "^4.3.1"
vin-decoder-nhtsa = "^0.0.2"
requests = "^2.31.0"
flake8 = "^6.1.0"
black = "^23.11.0"
mypy = "^1.7.1"
django-stubs = "^4.2.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
